Move validator into validator directory
