+++ tests/agent_features/test_notice_error.py
@@ -19,7 +19,6 @@ from testing_support.fixtures import (
     error_is_saved,
     override_application_settings,
     reset_core_stats_engine,
-    validate_transaction_error_event_count,
     validate_transaction_error_trace_count,
 )
 from testing_support.validators.validate_application_error_event_count import (
@@ -30,6 +29,9 @@ from testing_support.validators.validate_application_error_trace_count import (
 )
 from testing_support.validators.validate_application_errors import (
     validate_application_errors,
+)
+from testing_support.validators.validate_transaction_error_event_count import (
+    validate_transaction_error_event_count,
 )
 from testing_support.validators.validate_transaction_errors import (
     validate_transaction_errors,
