+++ tests/testing_support/validators/validate_transaction_error_event_count.py
@@ -1,0 +1,35 @@
+# Copyright 2010 New Relic, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from newrelic.common.object_wrapper import transient_function_wrapper
+
+
+def validate_transaction_error_event_count(num_errors=1):
+    """Validate that the correct number of error events are saved to StatsEngine
+    after a transaction
+    """
+
+    @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
+    def _validate_error_event_on_stats_engine(wrapped, instance, args, kwargs):
+        try:
+            result = wrapped(*args, **kwargs)
+        except:
+            raise
+        else:
+            error_events = list(instance.error_events)
+            assert len(error_events) == num_errors
+
+        return result
+
+    return _validate_error_event_on_stats_engine
